Feat: improve org-ai-insert-snippet related part
The name of old function `org-ai-insert-snippets` can be misunderstood
because it just load the snippets to yasnippet collection.Also, it should not
be `interactive`, user can be suggested to load that function in their
config file only 1 time during startup. This commit change this function
name to `org-ai--load-snippets`

Also, this commit add new command org-ai-insert-snippet (snippet not
snippets like old name), it act the same as yasnippet-insert-snippet
command, that mean an interactive snippet picker will pop up and user
will pick one of them to insert to current location. The different is
that this function only filter and load those snippet who match (or
contain) the string (or may be regex) custom variable name
`org-ai-snippet-regex`  (which default is "org-ai")