Feature: add option to save images as attachments (#140)
This adds the option `org-ai-save-image-as-attachment` to store images
as attachments of the current org heading instead in a static dir.

This PR also makes the filename more unique by adding the time of day.
This is necessary to prevent collision, because the files are moved to
the attachment dir after downloading them to `org-ai-image-directory` (I
considered downloading directly to attachment dir but it would be less
canonical and implies more code changes). I also ensured that images
created by the same request have the same timestamp.