Add `ai` block to `org-structure-template-alist` (#148)
Just a minor convenience addition. With this, users can utilize
`org-insert-structure-template` to insert AI blocks and they can also
use `<A` completion with org-tempo. Also it is useful to register AI
blocks so that other packages like
[corg.el](https://github.com/isamert/corg.el) can use this info to
provide completion etc. (it does not work yet but I'm thinking of adding
such completions to corg.el)